Read the EXT1, EXT2 and EXT3 operand-width prefixes
A prefix widens the next instruction's first operand, its second, or both.
Skipping one and reading the instruction that follows at the narrow width
misreads every byte after it, so a chunk with more than 255 registers or
symbols was decoded into a different program: a 300-symbol source reached
ARYPUSH, an opcode it never contained, and a 250-local one reached a byte
that is not an opcode at all.

fetch_next consumes the prefix as part of the instruction it belongs to,
following the FETCH_*_1/_2/_3 rules in mruby's include/mruby/opcode.h. The
operand table becomes a table of shapes so one fetch serves every width, and
Fetched carries u16 where an operand can be widened. Registers grow past 256
on demand, because a frame that needs EXT1 needs more registers than the
fixed array had.

+++ mrubyedge/src/yamrb/optable.rs
@@ -1092,9 +1092,9 @@ pub(crate) fn do_op_send(
     vm: &mut VM,
     recv_index: usize,
     blk_index: Option<usize>,
-    a: u8,
-    b: u8,
-    c: u8,
+    a: u16,
+    b: u16,
+    c: u16,
 ) -> Result<(), Error> {
     let method_id = vm.current_irep.syms[b as usize].clone();
     do_op_send_with_id(vm, recv_index, blk_index, a, method_id, c)
@@ -1104,9 +1104,9 @@ pub(crate) fn do_op_send_with_id(
     vm: &mut VM,
     recv_index: usize,
     blk_index: Option<usize>,
-    a: u8,
+    a: u16,
     method_id: RSym,
-    c: u8,
+    c: u16,
 ) -> Result<(), Error> {
     let mut n: usize = (c & 0x0f) as usize;
     let k: usize = (c >> 4) as usize;
